requirements.txt
streamlit
pandas
matplotlib
openpyxl
python-docx
chromadb
tornado==6.1